Move Repository.kt to data:repository module.
