+++ app/build.gradle.kts
@@ -112,6 +112,7 @@ dependencies {
     implementation(project(":core"))
     implementation(project(":model"))
     implementation(project(":data:remote"))
+    implementation(project(":data:repository"))
 
     implementation(Deps.Kotlin.stdlib)
     implementation(Deps.Kotlin.reflect)
