+++ app/src/main/java/jp/kuaddo/tsuidezake/ui/drink/DrinkDetailViewModel.kt
@@ -6,7 +6,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.liveData
 import com.squareup.inject.assisted.Assisted
 import com.squareup.inject.assisted.AssistedInject
-import jp.kuaddo.tsuidezake.data.Repository
+import jp.kuaddo.tsuidezake.data.repository.Repository
 import jp.kuaddo.tsuidezake.delegate.SnackbarViewModelDelegate
 import jp.kuaddo.tsuidezake.model.ErrorResource
 import jp.kuaddo.tsuidezake.model.SakeDetail
